:recycle: refactor: (App) 컨트롤러 역할을 변경 후 UI로직으로 앱 구성 UI로직을 앱이 수행하도록 게임 컨트롤러를 새로 추가해서 모듈을 연결함
